test: split tests into their own files
